Move Language Provider store to central store
