+++ src/containers/LanguageProvider/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { useSelector } from 'react-redux';
 import { IntlProvider } from 'react-intl';
-import { makeSelectLocale } from './selectors';
+import { makeSelectLocale } from 'store/language/selectors';
 
 export function LanguageProvider({ messages, children }) {
   const locale = useSelector(makeSelectLocale());
@@ -16,7 +16,7 @@ export function LanguageProvider({ messages, children }) {
 
 LanguageProvider.propTypes = {
   messages: PropTypes.object,
-  children: PropTypes.element.isRequired
+  children: PropTypes.element.isRequired,
 };
 
 export default LanguageProvider;
