+++ src/store/language/actions.js
@@ -1,8 +1,8 @@
-import { CHANGE_LOCALE } from './constants';
+import { CHANGE_LOCALE } from './actionTypes';
 
 export function changeLocale(languageLocale) {
   return {
     type: CHANGE_LOCALE,
-    locale: languageLocale
+    locale: languageLocale,
   };
 }
