+++ src/store/language/reducer.js
@@ -1,14 +1,14 @@
 import produce from 'immer';
-import { CHANGE_LOCALE } from './constants';
+import { CHANGE_LOCALE } from './actionTypes';
 import { DEFAULT_LOCALE } from '../../i18n';
 
 export const initialState = {
-  locale: DEFAULT_LOCALE
+  locale: DEFAULT_LOCALE,
 };
 
 /* eslint-disable default-case */
 const languageReducer = (state = initialState, action) =>
-  produce(state, draft => {
+  produce(state, (draft) => {
     switch (action.type) {
       case CHANGE_LOCALE:
         draft.locale = action.locale;
