+++ src/store/reducers.js
@@ -4,11 +4,11 @@ import { connectRouter } from 'connected-react-router';
 import history from 'utils/history';
 import loadingReducer from './loading/reducer';
 import authReducer from './auth/reducer';
-import languageProviderReducer from 'containers/LanguageProvider/reducer';
+import languageReducer from './language/reducer';
 
 export default function createReducer(injectedReducers = {}) {
   const rootReducer = combineReducers({
-    language: languageProviderReducer,
+    language: languageReducer,
     router: connectRouter(history),
     loading: loadingReducer,
     auth: authReducer,
